% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2correlate_samples.R
\name{r2correlate_samples}
\alias{r2correlate_samples}
\title{Correlate Aggregated Expression Data from a Seurat Object}
\usage{
r2correlate_samples(
  seurat_obj,
  group_by = "orig.ident",
  assay = "RNA",
  return_heatmap = TRUE
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the single-cell RNA-seq data.}

\item{group_by}{A character string specifying the metadata column in the Seurat object to group cells by.
This can be a column like `"orig.ident"` or any other metadata column in `seurat_obj@meta.data`.}

\item{assay}{A character string indicating the assay to use from the Seurat object. Default is `"RNA"`.}

\item{return_heatmap}{A logical value. If `TRUE`, the function returns a heatmap of the correlation matrix using the `pheatmap` package.
If `FALSE`, the function returns the correlation matrix itself. Default is `TRUE`.}
}
\value{
If `return_heatmap = TRUE`, returns a heatmap of the correlation matrix.
  If `return_heatmap = FALSE`, returns a matrix where each element represents the correlation between
  two groups' aggregated gene expression.
}
\description{
This function aggregates gene expression data from a Seurat object based on a specified grouping factor
(e.g., sample or cluster) and computes the correlation matrix between the groups. Optionally, it can return
a heatmap visualization of the correlation matrix.
}
\details{
The function aggregates gene expression data based on the `group_by` parameter using the `AggregateExpression()`
function from Seurat. It then computes the correlation matrix between groups using Pearson's correlation
(`cor` function with `use = "complete.obs"` to handle missing values). If requested, the function will visualize the
correlation matrix as a heatmap using the `pheatmap` package.
}
\examples{
\dontrun{
# Example usage:
pbmc <- CreateSeuratObject(pbmc_data)
correlation_matrix <- r2correlate_samples(seurat_obj = pbmc, group_by = "orig.ident", assay = "RNA", return_heatmap = FALSE)

# Return a heatmap instead of the matrix
r2correlate_samples(seurat_obj = pbmc, group_by = "orig.ident", return_heatmap = TRUE)
}

}
