% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2easyviz-package.R
\docType{package}
\name{R2easyviz-package}
\alias{R2easyviz}
\alias{R2easyviz-package}
\title{R2easyviz: Quick and Easy Seurat-Compatible Plots}
\description{
Provides a host of functions to quickly and easily make Seurat-compatible plots. Choose your Seurat object and the metadata of interest, and plot away.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryanreis333/R2easyviz}
  \item Report bugs at \url{https://github.com/ryanreis333/R2easyviz/issues}
}

}
\author{
\strong{Maintainer}: Ryan Reis \email{ryanreis333@gmail.com}

}
\keyword{internal}
