% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2heatmap.R
\name{r2heatmap}
\alias{r2heatmap}
\title{Generate a Heatmap of Marker Genes}
\usage{
r2heatmap(
  seurat_obj,
  FindAllMarkersObj,
  group_by = "celltype",
  ncells = 500,
  viridis_color = TRUE,
  nfeatures = 5,
  arrange_by = "dif",
  group_colors = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{FindAllMarkersObj}{A data frame from `FindAllMarkers` containing marker genes.}

\item{group_by}{A metadata column to group cells by in the heatmap. Default is `"celltype"`.}

\item{ncells}{An integer for the number of cells to sample per group. Default is 500.}

\item{viridis_color}{A logical indicating whether to use the Viridis color scale. Default is `TRUE`.}

\item{nfeatures}{An integer for the number of top marker genes per cluster to show. Default is 5.}

\item{arrange_by}{A column name in `FindAllMarkersObj` to order genes by.
If `"dif"`, it's calculated as `pct.1 - pct.2`. Default is `"dif"`.}

\item{group_colors}{An optional named vector for custom group colors.}
}
\value{
A `ggplot` object representing the heatmap.
}
\description{
The `r2heatmap` function creates a heatmap of top marker genes from a Seurat object.
It visualizes the expression of selected markers across a downsampled set of cells,
with options for custom colors and grouping.
}
\examples{
\dontrun{
# To create a dummy Seurat object for testing
pbmc_data <- matrix(rnorm(1000 * 20), nrow = 1000)
rownames(pbmc_data) <- paste0("Gene-", 1:1000)
colnames(pbmc_data) <- paste0("Cell-", 1:20)
seurat_obj <- CreateSeuratObject(counts = pbmc_data)
seurat_obj$celltype <- sample(c("T-cell", "B-cell"), 20, replace = TRUE)
Idents(seurat_obj) <- seurat_obj$celltype
markers_df <- FindAllMarkers(seurat_obj, only.pos = TRUE, min.pct = 0.25, logfc.threshold = 0.25)

# Generate a heatmap
r2heatmap(seurat_obj, FindAllMarkersObj = markers_df, group_by = "celltype")
}

}
