% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2prop_by_group.R
\name{r2prop_by_group}
\alias{r2prop_by_group}
\title{Grouped Bar Plot with Individual Sample Points for Seurat Object}
\usage{
r2prop_by_group(
  seurat_object,
  celltype = Idents(seurat_object),
  group = "group",
  orig.ident = "orig.ident"
)
}
\arguments{
\item{seurat_object}{A Seurat object containing the data to be plotted.}

\item{celltype}{A string specifying the column name for cell types.}

\item{group}{A string specifying the column name for grouping.}

\item{orig.ident}{A string specifying the column name for sample identifiers.}
}
\value{
A ggplot object that visualizes the grouped bar plot with individual sample points.
}
\description{
Grouped Bar Plot with Individual Sample Points for Seurat Object
}
