% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2propplot.R
\name{r2propplot}
\alias{r2propplot}
\title{Title}
\usage{
r2propplot(
  seurat_object = "so",
  celltype = "celltype",
  orig.ident = "orig.ident"
)
}
\arguments{
\item{seurat_object}{A seurat object you want to create the plot from}

\item{celltype}{The celltype or other metadata you want to calculate proportions from}

\item{orig.ident}{The orig.ident or sample name you want to split the data by}
}
\value{
A ggplot that shows the proportion of each celltype across samples
}
\description{
Title
}
\examples{
r2propplot()

}
