% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2sample_seurat.R
\name{r2sample_seurat}
\alias{r2sample_seurat}
\title{Downsample Seurat Object by Group}
\usage{
r2sample_seurat(seurat_obj, group_var = "orig.ident", n = 500)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the single-cell RNA-seq data.}

\item{group_var}{A character string representing the metadata column used to define the groups for sampling. Default is `"orig.ident"`.}

\item{n}{An integer specifying the number of cells to sample from each group. Default is 500.}
}
\value{
A downsampled Seurat object containing the specified number of cells from each group.
}
\description{
`r2sample_seurat` is designed to downsample a Seurat object by randomly sampling
a specified number of cells (`n`) from each group defined by a metadata variable (`group_var`).
The function returns a new Seurat object containing only the sampled cells.
}
\examples{
\dontrun{
# To create a dummy Seurat object for testing
pbmc_data <- matrix(rnorm(1000 * 20), nrow = 1000)
rownames(pbmc_data) <- paste0("Gene-", 1:1000)
colnames(pbmc_data) <- paste0("Cell-", 1:20)
seurat_obj <- CreateSeuratObject(counts = pbmc_data)
seurat_obj$orig.ident <- sample(c("SampleA", "SampleB"), 20, replace = TRUE)

# Downsample 10 cells from each sample group defined by `orig.ident`
pbmc_sampled <- r2sample_seurat(seurat_obj = seurat_obj, group_var = "orig.ident", n = 10)
}

}
